+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".Skourse_Search"></activity>
         <activity
             android:name=".MainActivity"
-            android:label="@string/title_activity_skourse__search"></activity>
-        <activity android:name=".Skourse_Search">
+            android:label="@string/title_activity_skourse__search" />
+        <activity android:name=".Skourse_Main">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
